% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entrez2fantom}
\alias{entrez2fantom}
\title{ENTREZ IDs to FANTOM5 core promoters names mappings}
\format{
A named \code{character} vector of length 14214.
}
\usage{
entrez2fantom
}
\description{
Vector mapping ENTREZ IDs to FANTOM5 core promoters names.
}
\examples{
entrez2fantom()

}
\keyword{datasets}
