% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoters_f5_core}
\alias{promoters_f5_core}
\title{Core promoters GenomicRanges object}
\format{
A \code{GenomicRanges} object of length 14191, with 16 metadata columns:
\describe{
  \item{name}{Promotor name.}
  \item{score}{Numeric vector.}
  \item{gene_type_gencode}{Gene type of associated gene as defined by GENCODE annotation.}
  \item{ENTREZID}{ENTREZ ID of associated gene as defined by all three annotation sources.
                  Prevalence of annotations: UCSC > GENCODE > FANTOM5.}
  \item{SYMBOL}{Gene symbol of associated gene as defined by all three annotation sources.
                Prevalence of annotations: UCSC > GENCODE > FANTOM5.}
}
}
\usage{
promoters_f5_core
}
\description{
Core promoters selected from promoters_f5.
Selection criteria were GENCODE confirmation and ENCODE ROADMAP confirmation.
Further for each gene single promoter with highest FANTOM5 score was selected.
}
\examples{
promoters_f5_core()

}
\keyword{datasets}
