% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{symbol2fantom}
\alias{symbol2fantom}
\title{Gene symbols to FANTOM5 core promoters names mappings}
\format{
A named \code{character} vector of length 14222.
}
\usage{
symbol2fantom
}
\description{
Vector mapping gene symbols to FANTOM5 core promoters names.
}
\examples{
symbol2fantom()

}
\keyword{datasets}
